Consistently use the same clock in campaigns
